+++ src/app/views/OnboardingFollows.svelte
@@ -26,7 +26,7 @@
 
   const {load: loadPeople} = createPeopleLoader()
 
-  const prev = () => setStage("profile")
+  const prev = () => setStage("keys")
 
   const next = async () => {
     loading = true
@@ -119,7 +119,7 @@
 <div class="flex gap-3">
   <p
     class="-ml-1 -mt-2 flex h-12 w-12 items-center justify-center rounded-full bg-neutral-700 text-lg">
-    4/5
+    3/4
   </p>
   <p class="text-2xl font-bold">Find your people</p>
 </div>